Move TBillingStats to schemeshard_billing_helpers (#19658)
